backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.0
pytest-asyncio==0.23.0
